autosettle: keep autosettler active until actually building a city
This is to fix the situation where several settlers all pick the
same "ideal location" for their new city, and they all set out for
the same place. One of the settlers arrives first and builds the
city, the rest get there and try to perform a "build-city" command
but fail because there is a city already there.